requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
apscheduler>=3.10.4
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
